+++ src/wacfrac/orbit.cpp
@@ -7,7 +7,7 @@
 namespace wacfrac
 {
 
-auto rebase_reference(const std::vector<std::complex<double>>& ref, std::size_t ref_n, std::complex<double> dz) -> std::tuple<std::size_t, std::complex<double>, std::complex<double>> {
+auto rebase_reference(const std::vector<std::complex<long double>>& ref, std::size_t ref_n, std::complex<long double> dz) -> std::tuple<std::size_t, std::complex<long double>, std::complex<long double>> {
     if (ref_n >= ref.size()) {
         return {0uz, dz, dz};
     }
@@ -21,33 +21,33 @@ auto rebase_reference(const std::vector<std::complex<double>>& ref, std::size_t 
     return {ref_n, dz, z};
 }
 
-auto compute_next_perturbation(const std::vector<std::complex<double>>& ref, std::size_t ref_n, std::complex<double> dc, std::complex<double> dz) -> std::tuple<std::size_t, std::complex<double>, std::complex<double>> {
-    dz = 2.0 * dz * ref[ref_n] + dz * dz + dc;
+auto compute_next_perturbation(const std::vector<std::complex<long double>>& ref, std::size_t ref_n, std::complex<long double> dc, std::complex<long double> dz) -> std::tuple<std::size_t, std::complex<long double>, std::complex<long double>> {
+    dz = 2.0L * dz * ref[ref_n] + dz * dz + dc;
     ++ref_n;
     return rebase_reference(ref, ref_n, dz);
 }
 
 // https://fractalforums.org/index.php?topic=4360.0
-auto escape_perturbed(const std::vector<std::complex<double>>& ref, std::complex<double> dc, std::size_t max_n, std::complex<double> dz, std::size_t n) -> std::pair<std::complex<double>, std::size_t> {
+auto escape_perturbed(const std::vector<std::complex<long double>>& ref, std::complex<long double> dc, std::size_t max_n, std::complex<long double> dz, std::size_t n) -> std::pair<std::complex<long double>, std::size_t> {
     auto ref_n {n};
     if (ref_n >= ref.size()) {
         return {{}, 0};
     }
 
     auto z0 {ref[ref_n] + dz};
-    return escape_generic(z0, n, max_n, [&](std::complex<double> z) {
+    return escape_generic(z0, n, max_n, [&](std::complex<long double> z) {
         std::tie(ref_n, dz, z) = compute_next_perturbation(ref, ref_n, dc, dz);
         return z;
     });
 }
 
-auto compute_reference(multi_complex c, std::size_t max_n, bool do_escape) -> std::vector<std::complex<double>> {
-    std::vector<std::complex<double>> reference { {0.0, 0.0} };
+auto compute_reference(multi_complex c, std::size_t max_n, bool do_escape) -> std::vector<std::complex<long double>> {
+    std::vector<std::complex<long double>> reference { {0.0, 0.0} };
 
     multi_complex z {0.0, 0.0};
     for (auto n {0uz}; n < max_n - 1 && (!do_escape || !escaped(z)); ++n) {
         z = compute_next_z(z, c);
-        reference.emplace_back(static_cast<std::complex<double>>(z));
+        reference.emplace_back(static_cast<std::complex<long double>>(z));
     }
 
     return reference;
