+++ src/wacfrac/viewport.cpp
@@ -19,26 +19,26 @@ auto viewport::sample(std::size_t x, std::size_t y, std::size_t width, std::size
     return multi_complex(re, im);
 }
 
-auto viewport::generate_probes(std::size_t cols, std::size_t rows) const -> std::vector<std::complex<double>> {
-    std::vector<std::complex<double>> probes;
-    std::complex<double> interval {
-        static_cast<double>(dimensions.real()/cols),
-        static_cast<double>(dimensions.imag()/rows)
+auto viewport::generate_probes(std::size_t cols, std::size_t rows) const -> std::vector<std::complex<long double>> {
+    std::vector<std::complex<long double>> probes;
+    std::complex<long double> interval {
+        static_cast<long double>(dimensions.real()/cols),
+        static_cast<long double>(dimensions.imag()/rows)
     };
     if (interval.real() == 0.0 || interval.imag() == 0.0) {
         probes.emplace_back(0.0, 0.0);
         return probes;
     }
-    std::complex<double> start {
-        static_cast<double>(cols % 2 == 0 ? -dimensions.real()/2.0 : (-dimensions.real() + interval.real())/2.0),
-        static_cast<double>(rows % 2 == 0 ? -dimensions.imag()/2.0 : (-dimensions.imag() + interval.imag())/2.0)
+    std::complex<long double> start {
+        static_cast<long double>(cols % 2 == 0 ? -dimensions.real()/2.0 : (-dimensions.real() + interval.real())/2.0),
+        static_cast<long double>(rows % 2 == 0 ? -dimensions.imag()/2.0 : (-dimensions.imag() + interval.imag())/2.0)
     };
-    std::complex<double> end {
-        static_cast<double>(+dimensions.real()/2.0),
-        static_cast<double>(+dimensions.imag()/2.0)
+    std::complex<long double> end {
+        static_cast<long double>(+dimensions.real()/2.0),
+        static_cast<long double>(+dimensions.imag()/2.0)
     };
-    for (double dx = start.real(); dx <= end.real(); dx += interval.real()) {
-        for (double dy = start.imag(); dy <= end.imag(); dy += interval.imag()) {
+    for (long double dx = start.real(); dx <= end.real(); dx += interval.real()) {
+        for (long double dy = start.imag(); dy <= end.imag(); dy += interval.imag()) {
             probes.emplace_back(dx, dy);
         }
     }
